darkpool-client: transfer-auth: Move deposit flow to generic helper
