+++ darkpool-client/src/transfer_auth/base.rs
@@ -1,0 +1,1 @@
+//! Base implementation of transfer auth
